pyproject.toml
[project]
name = "mega-sena-analise"
version = "0.1.0"
description = ""
authors = [
    {name = "eriqueduarte",email = "eriquemcz@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pandas (>=2.3.3,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "beautifulsoup4 (>=4.14.2,<5.0.0)",
    "lxml (>=6.0.2,<7.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "python-telegram-bot (>=22.5,<23.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
